commit 6
ajout majeur et correction

+++ src/main/java/eni/tp/app/eni_app/DAO/IDAOMovie.java
@@ -8,6 +8,6 @@ public interface IDAOMovie {
 
     List<Movie> selectMovies();
 
-    Movie selectMovieById(long id);
+    Movie selectMovieById(int id);
 
 }
