+++ src/main/java/eni/tp/app/eni_app/bll/MovieManager.java
@@ -26,7 +26,7 @@ public class MovieManager {
 
     }
     public Movie getById(long id) {
-        Movie movie = daoMovie.selectMovieById(id);
+        Movie movie = daoMovie.selectMovieById((int) id);
         return movie;
     }
 }
